+++ infinite_test.ts
@@ -1,4 +1,4 @@
-import { count } from "./infinite.ts";
+import { count, cycle } from "./infinite.ts";
 import { assertStreamStartsWith } from "./testing.ts";
 
 Deno.test("count()", async () => {
@@ -7,4 +7,22 @@ Deno.test("count()", async () => {
   await assertStreamStartsWith(count(1, 3), [1, 4, 7, 10, 13, 16]);
   await assertStreamStartsWith(count(5, 1.5), [5, 6.5, 8, 9.5, 11, 12.5]);
   await assertStreamStartsWith(count(100, -1), [100, 99, 98, 97, 96, 95]);
+});
+
+Deno.test("cycle()", async () => {
+  await assertStreamStartsWith(
+    cycle(["foo", "bar", "baz"]),
+    ["foo", "bar", "baz", "foo", "bar", "baz", "foo", "bar", "baz", "foo"],
+  );
+
+  async function* gen() {
+    yield 3;
+    yield 6;
+    yield 9;
+  }
+
+  await assertStreamStartsWith(
+    cycle(gen()),
+    [3, 6, 9, 3, 6, 9, 3, 6, 9, 3, 6, 9, 3, 6, 9, 3, 6, 9, 3, 6, 9, 3, 6, 9, 3],
+  );
 });