+++ CHANGES.md
@@ -9,6 +9,7 @@ Version 0.2.0
 To be released.
 
  -  Added *mod.ts* module which re-exports everything from other modules.
+ -  Added `cycle()` function.
 
 
 Version 0.1.0